requirements.txt
streamlit
fpdf
qrcode
pillow